chore: 升级全部依赖至最新版本（Tailwind 4 / Vue 3.5 / Tauri 2.11 / rusqlite 0.40 等）
- Tailwind CSS 3 → 4：配置系统从 tailwind.config.ts 迁移至 CSS @theme
- lucide-vue-next → @lucide/vue（新包名）
- rusqlite 0.32 → 0.40：Connection::open → open_with_flags，u64 → i64
- Vue 3.5、Pinia 3、Vite 6、PrimeVue 4.5、Tauri 2.11 等全量升级
- vue-router 4、@tauri-apps/api 2.11 兼容升级

+++ src-tauri/src/platform/windows/usn.rs
@@ -159,7 +159,7 @@ impl FileEngine for FallbackFileEngine {
                     path: r.full_path.clone(),
                     name: r.file_name.clone(),
                     extension: r.extension.clone(),
-                    size: r.file_size,
+                    size: r.file_size as i64,
                     modified_at: r.last_write_time,
                     is_directory: r.is_directory,
                 });
